backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
python-jose>=3.3.0
python-multipart>=0.0.6
bcrypt>=4.0.1
supabase>=1.0.0
openai>=1.0.0
asyncio>=3.4.3
aiohttp>=3.8.4
pillow>=9.5.0
moviepy==1.0.3
openai-agents>=0.0.14
websockets>=11.0.0
redis>=5.0.0
replicate>=0.22.0
fal-client>=1.0.0
requests>=2.31.0
